+++ src/js/letters/components/Address.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import _ from 'lodash';
-import { set } from 'lodash/fp';
 
 import ErrorableSelect from './ErrorableSelect';
 import ErrorableTextInput from './ErrorableTextInput';
@@ -24,41 +23,6 @@ class Address extends React.Component {
 
   componentWillMount() {
     this.id = _.uniqueId('address-input-');
-  }
-
-  // TODO: Look into if this is the best way to update address,
-  // it is incredibly slow right now
-  handleChange = (fieldName, update) => {
-    let address = set(fieldName, update, this.props.value);
-    // if country is changing we should clear the state
-    if (fieldName === 'country') {
-      address = set('state', '', address);
-    }
-
-    // TODO: Run the validations with the new address type
-    this.props.onUserInput(this.inferAddressType(address));
-  }
-
-  isMilitaryCity = (city) => {
-    const lowerCity = city.toLowerCase().trim();
-
-    return lowerCity === 'apo' || lowerCity === 'fpo' || lowerCity === 'dpo';
-  }
-
-  /**
-   * Infers the address type from the address supplied and returns the address
-   *  with the "new" type.
-   */
-  inferAddressType = (address) => {
-    let type = addressTypes.domestic;
-    if (!['USA', 'US'].includes(address.country)) {
-      type = addressTypes.international;
-    } else if (address.militaryStateCode) {
-      // TODO: Make sure we clear this out if a state code is selected
-      type = addressTypes.military;
-    }
-
-    return Object.assign({}, address, { type });
   }
 
   render() {
@@ -110,13 +74,13 @@ class Address extends React.Component {
         options={stateList}
         value={this.props.value[stateField]}
         required={this.props.required}
-        onValueChange={(update) => this.handleChange(stateField, update)}/>)
+        onValueChange={(update) => this.props.onUserInput(stateField, update)}/>)
       : (<ErrorableTextInput label="State/province"
         name="province"
         autocomplete="address-level1"
         value={this.props.value[stateField]}
         required={false}
-        onValueChange={(update) => this.handleChange(stateField, update)}/>);
+        onValueChange={(update) => this.props.onUserInput(stateField, update)}/>);
 
     return (
       <div>
@@ -127,7 +91,7 @@ class Address extends React.Component {
           options={countries}
           value={selectedCountry}
           required={this.props.required}
-          onValueChange={(update) => this.handleChange('country', update)}/>
+          onValueChange={(update) => this.props.onUserInput('country', update)}/>
         <ErrorableTextInput errorMessage={errorMessages.addressOne}
           label="Street address"
           name="address"
@@ -135,21 +99,21 @@ class Address extends React.Component {
           charMax={30}
           value={this.props.value.addressOne}
           required={this.props.required}
-          onValueChange={(update) => this.handleChange('addressOne', update)}/>
+          onValueChange={(update) => this.props.onUserInput('addressOne', update)}/>
         <ErrorableTextInput
           label="Street address (optional)"
           name="address"
           autocomplete="street-address"
           charMax={30}
           value={this.props.value.addressTwo}
-          onValueChange={(update) => this.handleChange('addressTwo', update)}/>
+          onValueChange={(update) => this.props.onUserInput('addressTwo', update)}/>
         <ErrorableTextInput
           label="Street address (optional)"
           name="address"
           autocomplete="street-address"
           charMax={30}
           value={this.props.value.addressThree}
-          onValueChange={(update) => this.handleChange('addressThree', update)}/>
+          onValueChange={(update) => this.props.onUserInput('addressThree', update)}/>
         <ErrorableTextInput errorMessage={errorMessages.city}
           label={<span>City <em>(or APO/FPO/DPO)</em></span>}
           name="city"
@@ -157,7 +121,7 @@ class Address extends React.Component {
           charMax={30}
           value={this.props.value[cityField]}
           required={this.props.required}
-          onValueChange={(update) => this.handleChange(cityField, update)}/>
+          onValueChange={(update) => this.props.onUserInput(cityField, update)}/>
         {stateProvince}
         <ErrorableTextInput errorMessage={errorMessages.zipCode}
           additionalClass="usa-input-medium"
@@ -166,7 +130,7 @@ class Address extends React.Component {
           autocomplete="postal-code"
           value={this.props.value.zipCode}
           required={this.props.required}
-          onValueChange={(update) => this.handleChange('zipCode', update)}/>
+          onValueChange={(update) => this.props.onUserInput('zipCode', update)}/>
       </div>
     );
   }
